задание 5.14
